thermal: intel: hfi: Update data type of hfi_features::nr_table_pages
A Coverity static code scan raised a potential overflow_before_widen
warning when hfi_features::nr_table_pages is used as an argument to
memcpy in intel_hfi_process_event().

Even though the overflow can never happen (the maximum number of pages of
the HFI table is 0x10 and 0x10 << PAGE_SHIFT = 0x10000), using size_t as
the data type of hfi_features::nr_table_pages makes Coverity happy and
matches the data type of the argument 'size' of memcpy().

+++ drivers/thermal/intel/intel_hfi.c
@@ -142,7 +142,7 @@ struct hfi_instance {
  */
 struct hfi_features {
 	int		nr_classes;
-	unsigned int	nr_table_pages;
+	size_t		nr_table_pages;
 	unsigned int	cpu_stride;
 	unsigned int	class_stride;
 	unsigned int	hdr_size;
